Add files via upload
Weight Tracking Application (Java - Original Creator)

The goal of this application was to design a weight tracker that allowed for user login, register a new user, provide CRUD abilities on a database of the user's entered weights and dates, and provide an SMS message when user meets their goal weight. In order to meet these goals I needed to develop multiple layouts to include a screen for login, registering a new user, and a "goal reached" screen. The design of these screens were to be as efficient as possible while still maintaining all functionality. This goal was incorporating the database CRUD capabilities as an edit button within the recycler view of the user's "history". In order to maintain the efficiency of the application, all code was kept modular so that items such as button listener segments could be utilized in other portions of the code. The hardest portion of this application to code was the CRUD database functions. While I believe that my database setup was correct, the application of each function into other activities was difficult to apply. The CRUD operations were supposed to be housed with the "Edit" button on the recycler view, and from this view a popup window will appear that will give the user the option to edit segments of the cardview or delete the entry entirely. From the same history view screen the users should have the option to add a "New Entry" which will create an entry within the database. Seeing the difficulty of SQLite application within an android app, in the future this will be the starting point for future applicaitons.

I believe the most successful part of this design is the history view. By combining the recycler and card view layouts I was able to fit multiple functions into a single screen for the user. I believe this portion of the project demonstrates how you are able to fit numerous features within a limited space and still maintain a clean and organized look for the user.